requirements.txt
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.51.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-js-asset==2.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
django-rest-passwordreset
drf-yasg==1.21.4
flake8==6.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pyflakes==3.0.0
pytz==2022.6
requests==2.28.1
whitenoise==6.2.0
